Put move mechanis into a separate table
Now instead of a MoveChangelog, there is a MoveMechanic that has all the
mechanic data.
The session gets a default_version_group_id attribute that works just like
default_language_id, and is used for a Move.default_mechanic relation.
Move.power, Move.type, etc. are now association proxies to the
default_mechanic. Unfortulately, as association_proxy is not transparent
enough, this means querying and eagerloads using the versioned data will
need to be rewritten.